backend/requirements.txt
# PolyGrand Backend Requirements
# Python 3.12+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Algorand
algokit-utils==3.0.0
py-algorand-sdk==2.6.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.6.0

# Database (for future use)
sqlalchemy==2.0.0

# Real-time Communication
websockets==13.1

# Caching (for future use)
redis==5.2.0

# HTTP Client
httpx==0.27.0

# Environment Variables
python-dotenv==1.0.0

# File Upload Support
python-multipart==0.0.12
